Refactor compile recipe selection
