+++ test/units/99_multiroot/build.test.ts
@@ -3,8 +3,8 @@ import * as path from 'path'
 import * as sinon from 'sinon'
 import { lw } from '../../../src/lw'
 import { autoBuild } from '../../../src/compile/build'
-import { build, initialize } from '../../../src/compile/recipe'
-import * as recipe from '../../../src/compile/recipe'
+import { build, initialize } from '../../../src/compile/recipe-old'
+import * as recipe from '../../../src/compile/recipe-old'
 import { Events } from '../../../src/core/event'
 import { queue } from '../../../src/compile/queue'
 import { assert, get, mock, set } from '../utils'
